Add support for west-const for variables
West-const is `const int` whereas east-const is `int const`.
They mean the same thing and both can be present at the same time
(`const int const` means the same).

Global variables that are initialized do not seem to work for
knight-native failing with `Target label 1 is not valid` when running
`hex2`.