Add initial enum support
This allows enums to be parsed and their enumerators to be used like
constants.

Using enums as types is not yet supported.

Both C89-only no-trailing-commas and C99 trailing-commas are allowed.

We can always just treat enums as `int`s if we document it since
(C17 quoted):
```text
6.7.2.2:
The expression that defines the value of an enumeration constant
shall be an integer constant expression that has a value
representable as an int

6.7.2.3:
The identifiers in an enumerator list are declared as constants
that have type int and may appear wherever such are permitted. [...]

6.7.2.4:
Each enumerated type shall be compatible with char,
a signed integer type, or an unsigned integer
type. The choice of type is implementation-defined
but shall be capable of representing the
values of all the members of the enumeration.
```