Add support for forward declaration of structs
Forward declared structs are created with size = 0 and can not be
instantiated. After being defined the declared object is found and the
proper size is set. The same happens for typedefs.

Initially I just prepended a new object with the same name in the
`global_types` linked list which will always be found first since it's
before the declared type.
This didn't work for typedefs since the append to the global type list
and it seemed to break things when I changed it.