Add support for __DATE__ and __TIME__
For `__DATE__` the standard says (C17 6.10.8.1):
```text
The date of translation of the preprocessing translation unit: a character string literal of
the form "Mmm dd yyyy", where the names of the months are the same as those generated
by the asctime function, and the first character of dd is a space character if the value is
less than 10. If the date of translation is not available, an implementation-defined valid
date shall be supplied.
```

For `__TIME__` the standard says (C17 6.10.8.1):
```text
The time of translation of the preprocessing translation unit: a character string literal of
the form "hh:mm:ss" as in the time generated by the asctime function. If the time of
translation is not available, an implementation-defined valid time shall be supplied.
```

Since we don't have any way of getting the date or time we provide an
implementation defined valid time and date.