Add __STDC* macros
We set `__STDC_HOSTED__` since we always require a `main` and don't have
a way of turning off standard library headers.
We set `__STDC__` despite us not conforming fully since we're trying our
best and consider missing functionality to be "not implemented" rather
than "against design decisions".

We set `__STDC_VERSION__` to C89 because it's the standard with the
least amount of features and we can consider C99+ features to be
extensions to the standard.