Add run-pass test script and directory
This allows easier creation of smaller tests by just creating a single C
file that takes no arguments and must return 0 to pass.
All tests since 0032 have been of this type since they are easier to
create and modify.
The unnecessary copy and paste of boilerplate is annoying when creating
new tests.

This way it's also easier to give descriptive names for the tests.

They are run before test0000 since those tests include parts of the
standard library which makes codegen issues manifest inside code that
isn't explicitly in the test file. This makes it harder to debug.

The test file is a duplicate of test0040.